Switch to opensbi

+++ kernel/src/console.rs
@@ -9,7 +9,7 @@ struct Stdout;
 impl Write for Stdout {
     fn write_str(&mut self, s: &str) -> fmt::Result {
         for c in s.chars() {
-            console_putchar(c as usize);
+            console_putchar(c as i32);
         }
         Ok(())
     }
